Fix env level variables
The original design started out wrong. Variables that track versions
of components should be in common and this repo should be versioned
with tags such as "dev-1.2.3", "test-1.2.2", and "prod-1.2.1".
The purpose of the dev/test/prod environment directories should be
to manage things that are different between environment levels that
do NOT get promoted. For example, security setings in prod may be
more strict than the other environments.